+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
  - Check if `MODELTRANS_AVAILABLE_LANGUAGES` only contains strings.
+ - Make sure `settings.LANGUAGE_CODE` is never returned from `conf.get_available_languages()`
 
 ## 0.0.7 (2017-09-04)
  - Cleaned up the settings used by django-modeltrans [#19](https://github.com/zostera/django-modeltrans/pull/19).